move defaults and fix exports
