+++ src/js/reducers/filingPeriod.js
@@ -1,7 +1,8 @@
+import * as types from '../constants'
 /*
  * Set the default current filing period
  */
-export const filingPeriod = (state = '2017', action) => {
+export default (state = '2017', action) => {
   switch (action.type) {
   case UPDATE_FILING_PERIOD:
     return action.filingPeriod
