+++ src/js/reducers/institutions.js
@@ -1,9 +1,10 @@
+import * as types from '../constants'
 /*
  * Set isFetching to true when institutions are being requested
  * Set isFetching to false and populate the institutions key
  *   when data is received
  */
-export const institutions = (state = {}, action) => {
+export default (state = {}, action) => {
   switch (action.type) {
   case REQUEST_INSTITUTIONS:
     return {
